+++ packages/plugin-engages-ui/src/campaigns/containers/AutoAndManualForm.tsx
@@ -4,7 +4,7 @@ import {
   EmailTemplatesQueryResponse,
   IEngageMessageDoc,
   IEngageScheduleDate,
-  IIntegrationWithPhone
+  IIntegrationWithPhone,
 } from '@erxes/ui-engage/src/types';
 
 import { useLazyQuery } from '@apollo/client';
@@ -57,14 +57,14 @@ const AutoAndManualFormContainer = (props: FinalProps) => {
     emailTemplatesQuery,
     integrationsConfigsQuery,
     externalIntegrationsQuery,
-    integrationsQuery
+    integrationsQuery,
   } = props;
 
   const [
     clientPortalConfigsQuery,
-    { loading, data = {} as ClientPortalConfigsQueryResponse }
+    { loading, data = {} as ClientPortalConfigsQueryResponse },
   ] = useLazyQuery<ClientPortalConfigsQueryResponse>(
-    gql(clientPortalQueries.getConfigs)
+    gql(clientPortalQueries.getConfigs),
   );
 
   const handleClientPortalKindChange = useCallback(
@@ -72,7 +72,7 @@ const AutoAndManualFormContainer = (props: FinalProps) => {
       setBusinessPortalKind(businessPortalKind);
       clientPortalConfigsQuery({ variables: { kind: businessPortalKind } });
     },
-    [businessPortalKind]
+    [businessPortalKind],
   );
 
   const configs =
@@ -96,7 +96,7 @@ const AutoAndManualFormContainer = (props: FinalProps) => {
 
   for (const ext of externalIntegrations) {
     const locals = integrations.filter(
-      i => i._id === ext.erxesApiId && i.isActive
+      (i) => i._id === ext.erxesApiId && i.isActive,
     );
 
     for (const local of locals) {
@@ -104,7 +104,7 @@ const AutoAndManualFormContainer = (props: FinalProps) => {
         _id: local._id,
         name: local.name,
         phoneNumber: ext.telnyxPhoneNumber,
-        isActive: local.isActive
+        isActive: local.isActive,
       });
     }
   }
@@ -112,15 +112,15 @@ const AutoAndManualFormContainer = (props: FinalProps) => {
   const updatedProps = {
     ...props,
     templates: emailTemplatesQuery?.emailTemplates || [],
-    smsConfig: configs.find(i => i.code === 'TELNYX_API_KEY'),
+    smsConfig: configs.find((i) => i.code === 'TELNYX_API_KEY'),
     integrations: mappedIntegrations,
     clientPortalGetConfigs,
     businessPortalKind,
     handleClientPortalKindChange,
-    loading
+    loading,
   };
 
-  const content = formProps => (
+  const content = (formProps) => (
     <AutoAndManualForm {...updatedProps} {...formProps} />
   );
 
@@ -134,30 +134,30 @@ const withTemplatesQuery = withFormMutations<Props>(
         name: 'emailTemplatesQuery',
         options: ({ totalCountQuery }) => ({
           variables: {
-            perPage: totalCountQuery.emailTemplatesTotalCount
-          }
-        })
-      })
-    )(AutoAndManualFormContainer)
-  )
+            perPage: totalCountQuery.emailTemplatesTotalCount,
+          },
+        }),
+      }),
+    )(AutoAndManualFormContainer),
+  ),
 );
 
 let composers: any[] = [
   graphql(gql(queries.totalCount), {
-    name: 'totalCountQuery'
-  })
+    name: 'totalCountQuery',
+  }),
 ];
 
 const integrationEnabledQueries = [
   graphql(gql(integrationQueries.integrationsGetConfigs), {
-    name: 'integrationsConfigsQuery'
+    name: 'integrationsConfigsQuery',
   }),
   graphql(gql(integrationQueries.integrationsGetIntegrations), {
     name: 'externalIntegrationsQuery',
     options: () => ({
       variables: { kind: 'telnyx' },
-      fetchPolicy: 'network-only'
-    })
+      fetchPolicy: 'network-only',
+    }),
   }),
   graphql<Props, IntegrationsQueryResponse>(
     gql(integrationQueries.integrations),
@@ -166,11 +166,11 @@ const integrationEnabledQueries = [
       options: () => {
         return {
           variables: { kind: 'telnyx' },
-          fetchPolicy: 'network-only'
+          fetchPolicy: 'network-only',
         };
-      }
-    }
-  )
+      },
+    },
+  ),
 ];
 
 // if (isEnabled('clientportal')) {