+++ packages/plugin-mobinet-ui/src/modules/buildings/components/Row.tsx
@@ -1,13 +1,13 @@
 import { RowTitle } from '@erxes/ui-engage/src/styles';
 import ActionButtons from '@erxes/ui/src/components/ActionButtons';
 import Button from '@erxes/ui/src/components/Button';
-import ModalTrigger from '@erxes/ui/src/components/ModalTrigger';
 import Tip from '@erxes/ui/src/components/Tip';
 import { __ } from '@erxes/ui/src/utils/core';
 import React from 'react';
 import BuildingForm from '../containers/Form';
 import TextInfo from '@erxes/ui/src/components/TextInfo';
 import { IBuilding } from '../types';
+import { Icon, Label, ModalTrigger } from '@erxes/ui/src';
 
 type Props = {
   index: number;
@@ -77,15 +77,35 @@ const Row = (props: Props) => {
 
   const installationRequestIds = building.installationRequestIds || [];
   const ticketIds = building.ticketIds || [];
+  const onClick = () => {
+    props.history.push(`/mobinet/building/details/${building._id}`);
+  };
+
+  let statusText = 'Сүлжээ нэвтрээгүй';
+
+  switch (building.serviceStatus) {
+    case 'active':
+      statusText = 'Сүлжээ нэвтэрсэн';
+      break;
+    case 'inactive':
+      statusText = 'Сүлжээ нэвтрээгүй';
+      break;
+    case 'inprogress':
+      statusText = 'Нэвтрүүлэлт хийгдэж буй';
+      break;
+    default:
+      statusText = 'Сүлжээ нэвтрээгүй';
+      break;
+  }
 
   return (
     <tr>
       <td key={Math.random()}>
-        <RowTitle>{props.index}</RowTitle>
+        <RowTitle onClick={onClick}>{props.index}</RowTitle>
       </td>
 
       <td key={Math.random()}>
-        <RowTitle>{building.name || '-'}</RowTitle>
+        <RowTitle onClick={onClick}>{building.name || '-'}</RowTitle>
       </td>
 
       <td key={Math.random()}>
@@ -119,7 +139,11 @@ const Row = (props: Props) => {
       <td key={Math.random()}>
         <RowTitle>{building.networkType || '-'}</RowTitle>
       </td>
-
+      <td key={Math.random()}>
+        <Label lblColor={building.color} ignoreTrans={true}>
+          <span>{statusText}</span>
+        </Label>
+      </td>
       <td>
         <ActionButtons>
           {renderEditAction()}