+++ packages/plugin-mobinet-ui/src/modules/buildings/graphql/queries.ts
@@ -26,6 +26,7 @@ quarter {
 quarterId
 type
 networkType
+customFieldsData
 `;
 
 const commonParamsDef = `
@@ -39,6 +40,8 @@ $quarterId: String
 $searchValue: String
 $type: String
 $customQuery: JSON
+$networkType: NetworkType 
+$serviceStatus: ServiceStatus
 `;
 
 const commonParams = `
@@ -52,6 +55,8 @@ quarterId: $quarterId
 searchValue: $searchValue
 type: $type
 customQuery: $customQuery
+networkType: $networkType
+serviceStatus:$serviceStatus
 `;
 
 const listQuery = `
@@ -162,5 +167,5 @@ export default {
   buildingsByBoundsQuery,
   buildingsQuery,
   assets,
-  configs
+  configs,
 };