+++ pos/modules/checkout/hooks/useTDB.tsx
@@ -22,8 +22,8 @@ export const endPoint = (port?: string) =>
   (port ? `http://localhost:${port}` : TDB_DEFAULT_PATH) + `/ecrt1000`
 
 const useTDB = () => {
-  const paymentType = usePaymentType(BANK_CARD_TYPES.TDB)
-  return { paymentType }
+  const tdb = usePaymentType(BANK_CARD_TYPES.TDB)
+  return { paymentType: tdb }
 }
 
 export const useTDBTransaction = (options: {
@@ -31,7 +31,7 @@ export const useTDBTransaction = (options: {
   onError: () => void
 }) => {
   const { onCompleted, onError } = options
-  const paymentType = usePaymentType(BANK_CARD_TYPES.TDB)
+  const { paymentType } = useTDB()
   const { port } = paymentType?.config || {}
 
   const TDBTransaction = async (variables: { _id: string; amount: number }) => {
@@ -55,7 +55,10 @@ export const useTDBTransaction = (options: {
           })
           return !!onCompleted && onCompleted()
         }
-        toast({ description: `${ecrResult}`, variant: "destructive" })
+        toast({
+          description: `${JSON.stringify(ecrResult)}`,
+          variant: "destructive",
+        })
         !!onError && onError()
       })
       .catch((e) => {